+++ src/pluto_media/Table_Disc.cpp
@@ -299,7 +299,7 @@ if (is_null[1])
 return "NULL";
 
 char *buf = new char[511];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_ID.c_str(), (unsigned long) min(255,m_ID.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_ID.c_str(), (unsigned long) min((size_t)255,m_ID.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -404,7 +404,7 @@ if (is_null[9])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
