+++ src/pluto_media/Table_File.cpp
@@ -308,7 +308,7 @@ if (is_null[2])
 return "NULL";
 
 char *buf = new char[301];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Path.c_str(), (unsigned long) min(150,m_Path.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Path.c_str(), (unsigned long) min((size_t)150,m_Path.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -322,7 +322,7 @@ if (is_null[3])
 return "NULL";
 
 char *buf = new char[201];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Filename.c_str(), (unsigned long) min(100,m_Filename.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Filename.c_str(), (unsigned long) min((size_t)100,m_Filename.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -414,7 +414,7 @@ if (is_null[10])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
