+++ src/pluto_media/Table_Picture.cpp
@@ -270,7 +270,7 @@ if (is_null[1])
 return "NULL";
 
 char *buf = new char[15];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Extension.c_str(), (unsigned long) min(7,m_Extension.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Extension.c_str(), (unsigned long) min((size_t)7,m_Extension.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -284,7 +284,7 @@ if (is_null[2])
 return "NULL";
 
 char *buf = new char[511];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_URL.c_str(), (unsigned long) min(255,m_URL.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_URL.c_str(), (unsigned long) min((size_t)255,m_URL.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -350,7 +350,7 @@ if (is_null[7])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
